cambio de idioma del estado optimizacion de la actualizacion de imagen
